notes-client/pyproject.toml
[project]
name = "notes-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.13.1",
    "openai>=1.106.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
]
